+++ lib/src/configuration.dart
@@ -36,11 +36,17 @@ class Configuration {
   RealmSchema get schema => _schema;
 
   /// Creates a [Configuration] with schema objects for opening a [Realm].
-  Configuration(List<SchemaObject> schemaObjects)
+  /// [readOnly] controls whether a [Realm] is opened as readonly.
+  /// This allows opening it from locked locations such as resources,
+  /// bundled with an application.  The realm file must already exists.
+  Configuration(List<SchemaObject> schemaObjects, {bool readOnly = false})
       : _schema = RealmSchema(schemaObjects),
         _handle = realmCore.createConfig() {
     schemaVersion = 0;
     path = defaultPath;
+    if (readOnly) {
+      isReadOnly = true;
+    }
     realmCore.setSchema(this);
   }
 
@@ -53,7 +59,7 @@ class Configuration {
   }
 
   /// The platform dependent path to the default realm file - `default.realm`.
-  /// 
+  ///
   /// If set it should contain the name of the realm file. Ex. /mypath/myrealm.realm
   static String get defaultPath => _defaultPath ??= _initDefaultPath();
   static set defaultPath(String value) => _defaultPath = value;
@@ -83,10 +89,18 @@ class Configuration {
   /// If omitted the [defaultPath] for the platform will be used.
   String get path => realmCore.getConfigPath(this);
   set path(String value) => realmCore.setConfigPath(this, value);
+
+  /// Gets or sets a value indicating whether a [Realm] is opened as readonly.
+  /// This allows opening it from locked locations such as resources,
+  /// bundled with an application.
+  ///
+  /// The realm file must already exists at [path]
+  bool get isReadOnly => realmCore.getConfigReadOnly(this);
+  set isReadOnly(bool value) => realmCore.setConfigReadOnly(this, value);
 }
 
 /// A collection of properties describing the underlying schema of a [RealmObject].
-/// 
+///
 /// {@category Configuration}
 class SchemaObject {
   /// Schema object type.
@@ -103,7 +117,7 @@ class SchemaObject {
 }
 
 /// Describes the complete set of classes which may be stored in a `Realm`
-/// 
+///
 /// {@category Configuration}
 class RealmSchema extends Iterable<SchemaObject> {
   ///@nodoc