+++ lib/src/native/realm_core.dart
@@ -155,6 +155,16 @@ class _RealmCore {
     _realmLib.realm_config_set_schema_version(config.handle._pointer, version);
   }
 
+  bool getConfigReadOnly(Configuration config) {
+    int mode = _realmLib.realm_config_get_schema_mode(config.handle._pointer);
+    return mode == realm_schema_mode.RLM_SCHEMA_MODE_READ_ONLY;
+  }
+
+  void setConfigReadOnly(Configuration config, bool value) {
+    int mode = value ? realm_schema_mode.RLM_SCHEMA_MODE_READ_ONLY : realm_schema_mode.RLM_SCHEMA_MODE_AUTOMATIC;
+    _realmLib.realm_config_set_schema_mode(config.handle._pointer, mode);
+  }
+
   ConfigHandle createConfig() {
     final configPtr = _realmLib.realm_config_new();
     return ConfigHandle._(configPtr);